move input, usage to own file
